Consolidated action server definitions into a single file
