+++ ow_lander/scripts/grind_client.py
@@ -4,7 +4,7 @@
 # Research and Simulation can be found in README.md in the root directory of
 # this repository.
 
-from ow_lander.actions import arm
+from ow_lander import actions
 from ow_lander import node_helper
 from ow_lander import constants
 
@@ -32,5 +32,5 @@ parser.add_argument('ground_position', type=float, nargs='?',
 args = parser.parse_args()
 
 node_helper.call_single_use_action_client(
-  arm.GrindServer, **vars(args)
+  actions.GrindServer, **vars(args)
 )
