+++ ow_lander/scripts/light_set_intensity_server.py
@@ -4,7 +4,7 @@
 # Research and Simulation can be found in README.md in the root directory of
 # this repository.
 
-from ow_lander.actions import lander
+from ow_lander import actions
 from ow_lander import node_helper
 
-node_helper.spin_action_server(lander.LightSetIntensityServer)
+node_helper.spin_action_server(actions.LightSetIntensityServer)
